refactor(password): move password.ts to password subfolder
