+++ src/public/app/widgets/type_widgets/content_widget.ts
@@ -14,7 +14,7 @@ import CodeAutoReadOnlySizeOptions from "./options/code_notes/code_auto_read_onl
 import CodeMimeTypesOptions from "./options/code_notes/code_mime_types.js";
 import ImageOptions from "./options/images/images.js";
 import SpellcheckOptions from "./options/spellcheck.js";
-import PasswordOptions from "./options/password.js";
+import PasswordOptions from "./options/password/password.js";
 import EtapiOptions from "./options/etapi.js";
 import BackupOptions from "./options/backup.js";
 import SyncOptions from "./options/sync.js";
